Rename packages and reorganize classes
