+++ app/src/main/java/com/example/android/miniweather/MainActivity.kt
@@ -3,8 +3,7 @@ package com.example.android.miniweather
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 
-import com.example.android.miniweather.Manager.NavigationManager
-import com.example.android.miniweather.View.WeatherViewFragment
+import com.example.android.miniweather.ui.WeatherViewFragment
 
 class MainActivity : AppCompatActivity() {
