+++ app/src/main/java/com/example/android/miniweather/model/Condition.kt
@@ -1,4 +1,4 @@
-package com.example.android.miniweather.Models
+package com.example.android.miniweather.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
