+++ app/src/main/java/com/example/android/miniweather/model/Location.kt
@@ -1,4 +1,4 @@
-package com.example.android.miniweather.Models
+package com.example.android.miniweather.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
