+++ app/src/main/java/com/example/android/miniweather/presenter/WeatherViewContract.kt
@@ -1,8 +1,8 @@
-package com.example.android.miniweather.Presenter
+package com.example.android.miniweather.presenter
 
-import com.example.android.miniweather.Models.CityWeather
-import com.example.android.miniweather.Models.FavouriteCityModel
-import com.example.android.miniweather.Models.TemperatureUnit
+import com.example.android.miniweather.model.CityWeather
+import com.example.android.miniweather.model.FavouriteCityModel
+import com.example.android.miniweather.model.TemperatureUnit
 
 interface WeatherViewContract {
